Refactor to allow a Proc to act like an exec.Cmd
I had thought I could let the user pass in the exec.Cmd but that is
problematic in the Wait case so I just wrapped the exec.Cmd in order
to keep around the necessary state. Also, I found that you can't reuse
the exec.Cmd, so that also makes things simpler.