Remove default use of the environment for expansion
In addition to the documentation updates, this changes the default
behavior to prefer **not** using the parent process environment by
default and requiring an explicit `map[string]string` be used. I also
pointed out the helper functions that can make it easier to use the
parent environment, assuming that is what the user wants.

Fixes: #3